Pass module to instance by pointer
The instance object captures the reference to the module object. This is bad API design, because the owner of the module must remember to keep it alive as long as the instance object is used. Anyway, this change makes it only slightly better. Using temporary module objects is more difficult now, e.g. `instantiate(parse(wasm))` will not compile any more. Also, instance objects are now assignable. This is useful, because they are returned by value from instantiate().